Better validate whether an item is craftable
